Modularise and get things compiling on Java 11.
The StandardAdapter was trashed in the process, though.
