+++ src/main/java/org/asciidoclet/asciidoclet/DocletRenderer.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright 2013-2018 John Ericksen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.asciidoclet.asciidoclet;
+
+import com.sun.javadoc.Doc;
+
+/**
+ * Interface used to render a Javadoc Doc
+ *
+ * @author John Ericksen
+ */
+public interface DocletRenderer {
+
+    void renderDoc(Doc doc);
+}
