+++ src/main/java/org/asciidoclet/asciidoclet/StandardAdapter.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2015 John Ericksen
+ * Copyright 2013-2018 John Ericksen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.asciidoctor.asciidoclet;
+package org.asciidoclet.asciidoclet;
 
 import com.sun.javadoc.DocErrorReporter;
 import com.sun.javadoc.RootDoc;
-import com.sun.tools.doclets.standard.Standard;
 
 /**
  * Adapter class to use the Standard Javadoc Doclet in a non-static context.
@@ -27,14 +26,14 @@ import com.sun.tools.doclets.standard.Standard;
 public class StandardAdapter {
 
     public int optionLength(String option) {
-        return Standard.optionLength(option);
+        return 0;//Standard.optionLength(option);
     }
 
     public boolean start(RootDoc rootDoc) {
-        return Standard.start(rootDoc);
+        return true;//Standard.start(rootDoc);
     }
 
     public boolean validOptions(String[][] options, DocErrorReporter errorReporter) {
-        return Standard.validOptions(options, errorReporter);
+        return false;//Standard.validOptions(options, errorReporter);
     }
 }
