+++ src/main/java/org/asciidoclet/asciidoclet/Stylesheets.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2015 John Ericksen
+ * Copyright 2013-2018 John Ericksen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.asciidoctor.asciidoclet;
+package org.asciidoclet.asciidoclet;
 
 import com.google.common.io.Resources;
 import com.sun.javadoc.DocErrorReporter;
