+++ src/main/java/org/asciidoclet/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2015 John Ericksen
+ * Copyright 2013-2018 John Ericksen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,7 +33,7 @@
  *   <version>2.9</version>
  *   <configuration>
  *     <source>1.7</source>
- *     <doclet>org.asciidoctor.Asciidoclet</doclet>
+ *     <doclet>org.asciidoclet.Asciidoclet</doclet>
  *     <docletArtifact>
  *       <groupId>org.asciidoclet</groupId>
  *       <artifactId>asciidoclet</artifactId>
@@ -173,8 +173,8 @@
  *
  * @author https://github.com/johncarl81[John Ericksen]
  * @version 0.1.0
- * @see org.asciidoctor.Asciidoclet
+ * @see org.asciidoclet.Asciidoclet
  * @since 0.1.0
  * @serial (or @serialField or @serialData)
  */
-package org.asciidoctor;
+package org.asciidoclet;
