+++ src/test/java/org/asciidoclet/asciidoclet/StylesheetsTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2015 John Ericksen
+ * Copyright 2013-2018 John Ericksen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.asciidoctor.asciidoclet;
+package org.asciidoclet.asciidoclet;
 
 import com.sun.javadoc.DocErrorReporter;
+import org.asciidoclet.asciidoclet.DocletOptions;
+import org.asciidoclet.asciidoclet.Stylesheets;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.asciidoctor.asciidoclet.Stylesheets.JAVA6_STYLESHEET;
-import static org.asciidoctor.asciidoclet.Stylesheets.JAVA8_STYLESHEET;
-import static org.asciidoctor.asciidoclet.Stylesheets.JAVA9_STYLESHEET;
+import static org.asciidoclet.asciidoclet.Stylesheets.JAVA6_STYLESHEET;
+import static org.asciidoclet.asciidoclet.Stylesheets.JAVA8_STYLESHEET;
+import static org.asciidoclet.asciidoclet.Stylesheets.JAVA9_STYLESHEET;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.*;
 
@@ -33,7 +35,7 @@ public class StylesheetsTest {
     @Before
     public void setup() throws Exception {
         mockErrorReporter = mock(DocErrorReporter.class);
-        stylesheets = new Stylesheets(DocletOptions.NONE, mockErrorReporter);
+        stylesheets = new Stylesheets( DocletOptions.NONE, mockErrorReporter);
     }
 
     @Test
